requirements.txt
instaloader>=4.9.6
requests>=2.28.0
Pillow>=9.0.0 